Refactor android code. (#520)
