+++ src/python/platforms/android/__init__.py
@@ -20,7 +20,10 @@ from . import battery
 from . import constants
 from . import device
 from . import fetch_artifact
+from . import flash
 from . import gestures
 from . import logger
+from . import sanitizer
 from . import settings
+from . import ui
 from . import wifi
